Validate votes using config.points as a whitelist. Ref #3 (See notes)
This commit also rewrites the config file to include all point
values as strings. Really, once they are written to the DOM of the
client, they're traded back and forth as strings anyway. This lets
us validate by whitelisting (using indexOf) as well as include
point "values" that aren't numbers. In this case, I traded 100 for
a frowny face because in our process, if you even think something
might qualify as a 40, it's too big.